% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{doMain}
\alias{doMain}
\title{Core Batch Processing Workflow}
\usage{
doMain(lib_data, CONFIG)
}
\arguments{
\item{lib_data}{List. The output from the \code{prepLib} function.}

\item{CONFIG}{List. Global configuration containing mass tolerances and scoring weights.}
}
\description{
The primary engine of the package. It iterates through mzML files, performs
spectrum cleaning, identifies common peaks against the theoretical library,
filters by MS1/MS2 tolerance, and generates initial identification reports.
}
