% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{doProcess}
\alias{doProcess}
\title{Process Mass Spectrometry Data for Lipopeptides}
\usage{
doProcess(lib, CONFIG)
}
\arguments{
\item{lib}{List. The output from the \code{prepLib} function.}

\item{CONFIG}{List. Global configuration containing mass tolerances.}
}
\value{
A nested list containing processed data for each file, required for down-stream scoring.
}
\description{
First stage of the pipeline. Reads mzML files, performs peak picking and cleaning,
matches against the theoretical library, annotates fragments, and writes the MS2 detail CSV.
}
