% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{doReScore}
\alias{doReScore}
\title{Rescoring based on saved workspace RData}
\usage{
doReScore(rdata_path, CONFIG)
}
\arguments{
\item{rdata_path}{Character. Path to the _Scoring_Workspace.RData file.}

\item{CONFIG}{List. Global configuration containing new weights.}
}
\description{
Rescoring based on saved workspace RData
}
