% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{doScore}
\alias{doScore}
\title{Re-scoring and FDR Calculation}
\usage{
doScore(rdata_path, CONFIG)
}
\arguments{
\item{rdata_path}{Character. Path to the intermediate workspace file (e.g., "*_Scoring_Workspace.RData").}

\item{CONFIG}{List. Global configuration list. Weights provided here will override those used in the original \code{doMain} run.}
}
\description{
Performs a multi-dimensional scoring analysis on previously identified features
using saved intermediate data. It calculates the Final LipopepID Score and
determines the 1\% False Discovery Rate (FDR) threshold using a Target-Decoy approach.
}
