% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{doStat}
\alias{doStat}
\title{Statistical Summary and Result Aggregation}
\usage{
doStat()
}
\description{
Consolidates results across multiple processed samples. It performs
data cleaning (e.g., removing contradictory adducts), calculates abundance
statistics, and organizes 1\% FDR filtered results into dedicated directories.
}
