% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{prepLib}
\alias{prepLib}
\title{Prepare Theoretical Lipopeptide Library}
\usage{
prepLib(lib_path = "./final_lib.csv", CONFIG)
}
\arguments{
\item{lib_path}{Character. Path to the theoretical library CSV file.}

\item{CONFIG}{List. Global configuration containing adduct offsets and scoring parameters.}
}
\description{
This function processes a CSV file containing theoretical lipopeptide data,
expands it based on adduct offsets, handles unsaturation, and converts the
data into a format compatible with MSnbase Spectrum2 objects.
}
