% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{doScore}
\alias{doScore}
\title{Scoring and FDR Calculation with Memory-based Processed Data}
\usage{
doScore(process_list, lib, CONFIG, optimized_weights = NULL)
}
\arguments{
\item{process_list}{List. The output object returned by doProcess.}

\item{lib}{List. The output from the prepLib function.}

\item{CONFIG}{List. Global configuration.}

\item{optimized_weights}{List or Vector. Optional optimized weights from doWeight.}
}
\description{
Scoring and FDR Calculation with Memory-based Processed Data
}
